Fixed issue 41 - Add support for multi-screen applications
Added the "Remove Screen" button to the designer toolbar.

Show a tab bar (with one tab for each screen, scrollable if
necessary) in the designer.

Check that the form is Screen1 before setting the application icon
in the project settings when the icon property is changed.

Added blocks for opening another screen into the Control drawer. The
block genuses already existed, but until now they were not added to
any drawers.

Show the form name (along with the project name) in the blocks editor
title bar and header pane.

Modified YABlockCompiler so that the code generated for REPL includes
code that adjusts the screen name.

Modified Form.java so that closing the application works correctly
when there are multiple activities.