Fixed dependency checking for ant builds.
components/build.xml
Improved the dependency checking for the ai.apt task. This solves the problem
where changing a Java file in the annotations, common, or scripts package, or
changing a .txt file in scripts/templates did not cause the annotation
processor targets to be re-built.

build-common.xml:
Improved the dependency checking for the ai.javac task. This solves the problem
where changing a "public static final" variable in one Java file did not cause
another Java file that uses the variable to be recompiled.
Changed the ai.javac presetdef to a macrodef so it can run a <depend> task
before the <javac> task.

appengine/build.xml:
In the AiServerLib ai.javac task, exclude AiServerLib.jar from the classpath
to prevent the dependency checking from thinking that the code always needs to
be recompiled.

buildserver/build.xml:
In the BuildServer ai.javac task, exclude BuildServer.jar from the classpath
to prevent the dependency checking from thinking that the code always needs to
be recompiled.

JSONUtilTest.java:
Encoded all non-ascii characters as \uXXXX so that this file can be compiled
with javac option encoding=utf-8.