This change includes:
- Try to get more info on BlobWriteException and BlobReadException. I'm
not actually sure whether I've achieved that because I haven't been able
to cause the blob errors to happen in my server. I'm hoping that keeping
the underlying exception as part of the Blob exceptions will cause more
info to get printed.

- Make blob delete exceptions not be fatal, but log them as errors so
that we know if they're happening.

- Slip in a fix for uploading projects where we ignore an
android.keystore file in the project zip if it has one. This  makes it
easier to upload projects that are downloaded with the admin
functionality to download arbitrary user projects.