Added a shape property to the Button component.
Added shapes are:
- rounded: rectangle button with round corners
- rectangular: rectangular button with squared off corners
- oval: oval shaped button

If the user selects a Button other than default with the default color the
background no longer flashes yellow when clicked.  This was also true prior to
the shape property being added.  In another CL, I am planning on adding a fix so
that a Button flashes yellow when clicked no matter what shape or background
color it has.