Addresses issue # 231 -- modifies Texting component to work with Wifi.
This solution using Google Voice.  It requires users to sign up online for
Google Voice and to download and install the Google Voice mobile app.

The mobile app is used for incoming messages.  For outgoing messages
the app uses Google AccountManager to authenticate the user and then
uses their Google Voice account to send messages.

The revised component reuses the (slightly modified) ClientLoginHelper
that is used with the (deprecated) FusiontablesControl component. The
SendMessage function makes use of free software (as source code) found
here: http://code.google.com/p/google-voice-java/

It adds a new source file, SmsBroadcastReciever, which is started
statically in the manifest file and is configured to listen for
either Google Voice or built-in-SMS messages.  This file is modeled
on free software found here: https://github.com/cicada-dev.

9/26: Added permissions for accessing and managing user accounts,
which were available to REPL but not to packaged apps.

9/27: Modifed SmsBroadcastReceiver to start Texting component
if it is not already running.  The Texting component is still not
receiving the first incoming message.

10/2: Fixed the (9/27) issue of missing the first message by creating
an onInitializeListener interface to allow Form to notify components
when the Initialize block has been dispatched.

10/2: Added a static method to SmsBroadcastReceiver to search by
META-DATA, specified in the Manifest, for an App Inventor app that handles Sms.

10/3: Modified Texting component and BroadcastReceiver to notify the user
when the Texting app is not running.  Messages received when the app is not
running are cached. They are dispatched through MessageReceived when the app
resumes.  When the app is running, incoming messages are immediately dispatched.
This should be more or less consistent with the performance of the previous
version.

10/5: Modified Texting to use OAuth2Helper, same as used in FusiontablesControl.
10/11: Added code in Texting.onInitialized() to set isRunning true at that point
     because in REPL start up Form.onResume() is called before the Texting
     component registers to receive onReceive() callbacks.  Messages received
     by the running REPL app are now received normally.
10/12: Added conditions	 to Compiler.java to insert <META-DATA>	and <RECEIVER>
     tags into the Android Manifest only when Texting component	is used.
10/25: Final changes to YaVersion, etc. and fixed a missing jar bug in misc/launch-builserver
10/30: Simplified SmsBroadcastReceiver; now doesn't notify if receiving disables; supports multiple apps
11/4: Removed rendundant jars; created lib/oauth with jars from lib/fusiontables
11/5: Really removed redundant jars.
11/5: Changed cache to local storage and synchronized it.

Change-Id: I1f8b285107f85d9c300fc78137378c3a644c9628