Move the whitelist to the database to the App Engine Datastore.
Move the WhiteList to the datastore from a white deployed with the App
Engine Instance. This includes adding access to the App Engine standard
remote access api (also implemented in the "manhole" app) which will
be used by the whitelist update script. The sample whitelist, used
for when the white list was part of the deployment, has been removed.

Change-Id: I8bab244e1297bf606ea0aa24c3c378c9dfba3b12