Update Companion Error Reporting.
Arrange for block load errors to be reflected back as red triangles,
with messages, associated with the block that caused the error.

Runtime errors are now reported back to the blocks editor as well as
being displayed as a “toast” in the Companion. Note: Such reports are
throttled to once every 5 seconds to avoid a flood of errors when the
error occurs in an event handler that is called frequently (like from
the accelerometer).

Updated the Companion Version to 2.07nb7. Version 2.07nb6 is still
supported (with no new functionality). Older versions are no longer
supported and will result in an upgrade required warning.

Change-Id: I8308500a2c35df25730f3e4ceed9d23c51477eec