Fixup Reference Documentation
Fix references within the component documentation to not have “/learn”
in their URLs (we have deprecated it). Also convert the chrome on the
contents of reference/other. It isn’t clear that this documentation is
completely used (it still shows AI1 version blocks). The
tinywebdb.html is references from the component documentation, but the
rest may no longer be in use and should be deleted.

Change-Id: I8e55b58bbd56ca2fc8a4918fd26d5bbfaa927a64