Update Wording in the Multiple-Session Dialog Warning.
Make the dialog people see when they are logged in from more then one
location more user friendly. Also fix the forums link from the Help
menu to go directly to Google Groups instead of via the Website. This
is a work around for a Google problem (problems uploading images in
replys).

Change-Id: I66f6c3311fe5fc339d1ead2d11bc1498ef743e82