Refactoring ListPicker, ListView, and Spinner
Makes the implementation of the three components more maintainable by delegating
methods to the ElementsUtil class.

Fixes problems with setting positions in listview and spinner.

Deletes BeforePicking event from listview because it does not make sense (it
does in ListPicker before the user can change elements on the fly, but not in
listview, as elements are already showing).

Change-Id: I68cd2bdb2da76085a8ada4529e68ed51671addf8