Companion Upgrades and Cleanup
Arrange for the Companion, when used via the Emulator or USB to be
re-entrant. Now if the Companion is running and a person reloads their
browser, they can reconnect without restarting the Companion.

Removed the “aiphoneapp” (the old USB Repl from AI1). Removed vestiges
of AI1 from the Companion App. Rename some variables to be more
descriptive as well. Cleanup some logging to use Log.d instead of
calling System.err.println(). By using Log.d, we ensure that
information sent to the log is available in the ACRA data.

Change-Id: Ib462fe96aefd0f35af5595e4b3e9b0aba1980634