Fix two problems in runtime error dialog within Blocks Editor
Problem 1: As illustrated in Ghica's "Blocks disappearing" email, HTML from an
error message could end up being rendered in a runtime error dialog. This
commit property escapes the HTML so that it is not rendered.

Problem 2: As illustrated in Ghica's "Blocks disappearing" email, an error
message could be so large that the runtime error dialog could fill the
entire blocks workspace. This commit adds a "blocklyRuntimeErrorDialog"
CSS class to the runtime error dialog that limits the window to no more
than 75% of the width and height of the blocks workspace, and makes the
contents scrollable if it's too big.

Change-Id: I4150c11aa93f6b10d8cee657b30141671d2f9050