The App Inventor 2 Gallery.
This Commit implements an Gallery of Applications integrated into the
MIT App Inventor Server. It is currently disabled by default. To
enable, edit appengine-web.xml and set "use.gallery" to true. Note:
You will have to create a Google Cloud Storage bucket for the use of
the Gallery and set the value of "gallery.bucket" to point to it.

Change-Id: I54d7d7b5d979a0c4eada8f00b3ceb101d22b1629