Add a “Deobsfucate Text” block
The Deobsfucate text block takes an obsfucated string and deobsfucates
it. When you create the block and place it in the blocks workspace
you are prompted to the plain text string and the obsfucated version
is automatically added as an input to the deobsfucate block.

The primary purpose of this block is to help hide secret values such
as API keys for the Fusion Tables and Twitter components. It isn’t
foolproof, but makes it a little harder for someone to extract the key
from a packaged application.

Change-Id: Ibf55927a56b17029915a5ec52b47db34744b37b3