Added “IgnoreSslError” property to the WebViewer
When set (default is not set) the WebViewer can use SSL to access web
pages with a self signed certificate. Setting it exposes you to a Man
in the Middle Attack, but on the other hand it also lets you access
pages with a self signed certificate. The security trade-off needs to
be made by the application designer.

Change-Id: I960628ebb1ea0c7c3839b8aad5170ed32fef7229