Modify the Buildserver shutdown system
Modify the buildserver shutdown system to accept a delay
parameter. This parameter, specified in seconds, defines a grace
period between now and the specified time in the future. During the
grace period the health check function will return “unhealthy” while
the buildserver itself will continue to accept new jobs. This more
gracefully handles the time period between then we declare that we
want to shutdown the buildserver and when the load balancer finally
decides that this buildserver is truly unhealthy (which requires
several polls) and takes it out of service. During this grace period
any jobs given to the buildserver by the load balancer will still be
honored.

Change-Id: Icf9f3f5cf7c958ee57ed774ab6a27457e6b347dc