Combine Yail Forms sent to the Companion
When we have a queue of forms ready to go to the Companion, combine
them into one large Ajax request. Unfortunately we lose track of which
block each form is form. However combining forms is likely only going
to happen when this isn’t as important. Keeping track of blocks
requires a new support in the Companion, which is forthcoming.

Change-Id: I312729fda02155fc0b142e2149fab1bdbadb3921