Do not attempt to rename the “type” of a block.
Do not attempt to rename the type of a block when it’s name is
changed. Doing so can corrupt projects if the oldname of a block is a
substring of a type name (say component_set_get). If you create a
component named “et”. Make sure it has blocks in the workspace and
then rename it to “st” the project will become corrupted because
occurances of the string “component_set_get” will turn into
“component_sst_get.”

Change-Id: I2dbcb308a6f81f67a2ebdef645e70c4dfee93f7a