Update Splash Screen Management
Move the configuration of the Splash Screen into the Data Store. The
contents and size of the splash screen is fetched from the datastore
at runtime. The content can either be local, or a reference to an
iframe on another site.

Add an internal version number to the splash screen.  Provide a "Do
Not Show Again" Checkbox so that end-users can dismiss the splash
screen until the version is incremented.

Provide a python script "splashmanager.py" (in
appinventor/misc/splashscreen) to permit site administrators to update
the splash configuration. It requires the App Engine python SDK to be
installed.

Change-Id: I8b162fac35a1b3f2d0d78f1be6069f2d46c0f717