Add Help Menu Item to show the Splash Screen
Now that users can permanently dismiss the splash screen (until we
increment the version) there needs to be a way to explicitly decide to
see it again. This code adds a menu choice in the Help menu to do just
that. Also included is some minor code and comments cleanup.

Change-Id: Ica68d05678bd1748c1144a1e4c2e1820343e2c63