Ensure Image Manipulation doesn't use too much memory
When an image is smaller then the screen size, then scale it up by the
device pixel density. However when an image is larger then the screen,
sample it down to fit and then do not scale it again by pixel size.

The rationale for this change is that if someone uploads a “small”
image, they are likely to have sized it for their application, in
which case we will want to scale it based on device screen
density. However if the image is larger then the screen, we can
presume some other mechanism will be used to adjust its displayed size
(for example a set length and width in an Image component). In this
case we sample it down to fit on the screen but we do not scale it
based on pixel density.

Change-Id: I250ff444dd1f48780a06ae70ba61bf20a6f7e023