Add component name-type map for building apk
In ode code, component names are used to identify a component though
the variables storing component names appear to be "type". While
there's no harm in ode, here in build server, they need to be
separated. This method returns a name-type map, mapping the component
names used in ode to the corresponding type, aka fully qualified name.
The type will be used to build apk.