Arrange for a “Send Error Report” in the Companion
If ACRA is active, add a menu item, only on the Companion, offering to
send an error report. Once the report is sent, a two word report ID is
displayed to the user. We will ask people who submit reports to send
us an e-email (or update the forum) with the two word ID so we can
find their report in our database.

Change-Id: I9898b0077ce163bdbc6af9aae546fd833d9a8a6c