Set scaleFrameLayout scale to 1.0 when keyboard open
Set scaling to 1.0 when the keyboard is open. This effectively puts us
in Responsive mode when the keyboard is open. Set it back to the
appropriate value (when in Fixed mode) when the keyboard is
closed. This works around the issue where the screen gets improperly
scaled when the keyboard is open. We suspect there is a bad
interaction between our scaling in Fixed mode and the scaling that
Android provides (at least in Android 5.1.1) when the keybaord is
opened. So we remove ours.

Change-Id: I1cec18ea57ead5527885d96c8b395ab90c240f5d