Client Side Improvements and Bugfixes
Added support for ordering the components within a category of the
component palette in the designer.

Added support for renaming a method or property in blockly versioning.

Added support for allowing a component to add menu items to the
options menu (at runtime).

Added error message ERROR_FUNCTIONALITY_NOT_SUPPORTED_WIFI_DIRECT for
when a component needs Wi-Fi peer-to-peer connectivity, but the device
does not support it.

Reserved error numbers 2901-2999 for FIRST Tech Challenge in
ErrorMessages.java.

Added NxtPaletteHelper to show how to specify the order of components
in a category.

Fixed browser error that occurs if you change
AppInventorFeatures.allowMultiScreenApplications() to return false.

Change-Id: I5890295dcff89bd7d4197ec935b0986161f11ca1