Arrange for YoungAndroidFormUpgrader to ignore external components.
For now we ignore *all* components that we don’t find in the
COMPONENT_DATABASE when we are upgrading projects because we don’t
have a good way to determine if a given component is an external
component.

This change is important because without it when YaVersion is
incremented *all* projects with external components will fail to
load. When YaVersion is incremented, the upgrader is run on projects
saved with an old version. The presence of an external component
triggers an IllegalArugmentException in the upgrader because it cannot
find a system component to match. We inhibit that error in this
commit.

The correct solution will require some mechanism to handle upgrading
of external components. This is just a quick fix

Change-Id: I76df60dfbe100ca0e5bedf37f1e966c9910de78b