Patch to make sure event parameter blocks can be renamed
The English (aka internal) parameter names are stored in the
“eventparam” field of a block. However when a parameter block is
renamed (which can only happen when an event has multiple arguments
and you are renaming one to the others, say startX to startY) the
eventparam field was not being updated. This patch makes sure it is
updated, but is likely not being done “in the right place.” but until
we find the correct fix, this patch should alleviate the symptoms of
the problem.

Change-Id: I8c189b271a693793afbb8d9cf255b8f0672a70a8