Add Support for Multiple Dex Files
Add support for multiple classes.dex files in a packaged app. This
overcomes the limit of 65536 methods that otherwise would come into
play.

All normal App Inventor component code and up to 12 additional support
libraries are bundled into the primary classes.dex file. Additional
libraries are placed in classes2.dex. Android API 20 and higher
automatically utilizes multiple Dex files. Code from Google is
included here to support manually adding classes2.dex in order
versions (back to API 4, Tested to API 8).

The Companion is special cased, because it includes every possible
library, it takes a long time to start the first time it is run after
installation while the MultiDex code parses and optimizes the
classes2.dex file. In the Companion we do this optimization in a
background thread and display a Progress Dialog.

Change-Id: I5f05841f09cd03e96c421685f65098464aba1053