Permit isGCS in FileData to be null
Change the definition of FileData.isGCS from “boolean” (lower case
boolean) to a wrapped “Boolean” (upper case) because the later can
take on a null value. Older projects may not have a value stored for
isGCS, which means its value is effectively null. A null value cannot
be loaded into a FileData object is isGCS is a boolean because
booleans cannot be null. This results in an error when loading older
projects.

Change-Id: I653fef7f3141c0febbec837929d28e4a00b256ff