Make EditableProperty types changeable
Be able to update the “type” of an EditableProperty even after a
property is added. This is to facilitate the FirebaseToken property
which is persisted only if the FirebaseURL is non-default. So if the
App Inventor programmer changes the FirebaseURL (or sets it back to
the default) we have to change the type of the FirebaseToken property.

Change-Id: I3f6b39adf70e44b2df9e993304f815a2d4fd0d17