Adds the “Firebase” Component
The Firebase component provides a mechanism similar to
TinyWebDB. However it is hosted using a Firebase database
(firebase.com) and supports updates to variables from different copies
of an application. If multiple copies of an application are running on
different devices which are connected to the network. When one device
updates a variable the other will get a “DataChanged” event with the
new value. This permits a form of “cloud variable.”

The use of Firebase normally requires an account at firebase.com, but
by default an account sponsored by MIT is configured (via “branding”
commits which are not public) for the MIT sponsored service.

Author: Kristin Asmus <kna934@gmail.com>
Author: William S. Byrne <will2596@gmail.com>
Author: Jeffrey I. Schiller <jis@mit.edu>

Change-Id: I197445bd4ddd7eafb188c43569f902e7f5924c4d