Provide a Menu Item to Update the Companion
Provide a Menu Item under the “Help” Menu to trigger a manual update
of the Companion. Also fix a bug in the Companion itself so that the
correct CORS headers are returned during the update (also put a work
around in replmgr.js to deal with the non-fixed
Companion). Automatically disconnect the Companion after the update so
the user doesn’t have to use the “Reset Connection” item before
connecting to the newly installed companion.

Change-Id: I31215863163974cf8f23bb68b03200cad8229ae5