Add a flag to Companion Version
Add a flag to indicate that this version of the companion requires
Component Types to be fully qualified component names (includes the
package name). This is used by backwards compatibility code to
determine what type of Yail to send. This will permit us to release
the extension (cdk) code without requiring a “Flag Day” updates of
everyone’s Companion

Change-Id: I8bee33fbbed90d941688d47bbd63fa6fb2dc9a85