Add Compatibility Hack for Yail Generation
Add Code to generate non-fully-qualified component names when
connected to an older Companion from before the extension code was
added. This code is a bit hackish, but we can remove this code once we
deprecate the older companion. THIS COMMIT PROBABLY WANTS TO BE
SEPARATE FROM THE REST OF THE EXTENSIONS CODE SO WE CAN REVERT IT IN
THE FUTURE.

Change-Id: I819ca281ef4b2b51a0c814d065baccddcd0aaa5b