Removed unnecessary ftc res files.
Changed the behavior when a hardware component's DeviceName is not found in the
configuration. Previous to this change, it was a fatal error if an app had a
hardware component whose DeviceName was not found in the configuration and the
fatal error occurred just before the OpMode's Init event was triggered, even if
the OpMode didn't use that hardware component. With this change, the fatal error
will occur only when an OpMode actually tries to use a hardware component whose
DeviceName is not found in the configuration.

If an app has no FtcRobotController component, the Form will add the Exit menu
item to the options menu.

Added FtcGamepad Type property getter.

Updated to FTC SDK from 11/27/2015.

Updated lib/ftc with .aar files from ftctechnh/ftc_app beta release 20160104.

Added FtcRobotController.TelemetrySorted (property getter & setter).

Make designer change a FtcGamepad components's GamepadNumber property to "2" if
the name is FtcGamepad2 (localized).

Modified FtcHardwareDevice.initHardwareDevice() to check explicitly for hardware
being null rather than calling toString().

Corrected spelling of 'Analytics' in FTC LICENSE file.

Minor changes to block descriptions.

Updated to FTC SDK release v16.02.09.001.