Verify the Companion Package
Make sure that the package name of the Companion is matches this
version of App Inventor. Different versions of App Inventor (different
forks) should have their own Companions, each with its own package
name. This check makes sure that a person is using the correct
Companion for the version of App Inventor they are using. The expected
version is set in YaVersion. The default value provided is for the MIT
App Inventor service provided by MIT.

Change-Id: I23f48ab017020448a01a5a8d781f831fe6cec450