Add Support for Local Logins
This change adds support for local logins as well as maintaining
compatibility with Google authentication (when running on App Engine).

A “local login” is an account whose login name is an email address and
which has a password which we store in the UserData object (hashed via
PBKDF2). Two new flags “auth.usegoogle” and “auth.uselocal” control
whether or not Google Logins and/or Local logins are accepted.

If auth.usegoogle is false and auth.uselocal is true, a login page is
displayed which prompts for a person’s email address and password. A
“set/recovery password” link is also displayed which can be used to
change a person’s password or set a new password. This link is how local
login accounts are created. When clicked, the person is prompted for
their e-mail address. We then e-mail them a link which they click on to
confirm their account. This takes them to a page where they can set
their password.

Note: Because of restrictions imposed by Google, we cannot send e-mail
directly from App Engine. Instead we use a “helper” server which accepts
an HTTP POST operation from us (using the Urlfetch functionality of App
Engine). This helper does the actual sending of e-mail. This also
permits us to control where the email originates from (i.e., we can use
a server at MIT instead of a server at Google as some locations care
from where the mail comes). The source code for the “helper” server is
included in this commit (in the misc/passwordmail directory. See the
README file there for more details).

If auth.usegoogle and auth.uselocal are both set, the login page has a
link labeled “Use Google Account” which will then cause their Google
account to be used to login.

If auth.usegoogle is true and auth.uselocal is false, then the system
behaves as it does today. No local login page is shown, instead Google
accounts are used (which means that if someone isn’t logged in they see
Google’s account page as they do today).

We also add a “User Admin” feature that permits authorized users to
change the password (but not a Google password) for other users. Admins
can also “become” another user for purposes of examining or downloading
their projects. When in this “become” mode, the system is read-only. You
can look at other user’s project, and interact with them, but you cannot
save any work.

Change-Id: Id20c49b3dadbb5e5f2180364a928991dd31343bd