Add Atomic Functions to FirebaseDB
AppendValue is used to append a value to an existing list. It does it in
the context of a Firebase transaction so multiple clients can perform
this operation concurrently and no addition will be lost.

RemoveFirst takes the first element off of a list and returns it in the
FirstRemoved event. If two devices simultaneously invoke this, each will
get a different value. One will get the first value and the other will
get the second value.

If the tag provided is not a list, or the list is empty (for
RemoveFirst) an error is generated.

Also: If no FirebaseError event is defined by the App Inventor
Programmer, we put up a dialog box with the error. Previously in this
situations errors were silently ignored.

Change-Id: I7775a6d72d66d2a15df3366c5fd89a100b063a61