Update ActivityStarter to use TextUtils.isEmpty(String)
ActivityStarter uses String.isEmpty(), which was introduced in Java
1.6. However, older versions of Android are only binary compatible
with Java 1.5, such as the version used in the emulator. This causes
the ActivityStarter to throw an exception when starting an activity.

This commit fixes the issue by replacing the call to String.isEmpty()
with android.text.TextUtils.isEmpty(String), which is available since
Android API version 1. This is consistent with other checks in
components, such as Canvas and Web.

Closes #827

Change-Id: I5c26bab97421a86b3d62888af447d999af38450a