Pass the values of locale,repo and galleryId
“repo” is used by the template code to indicate a user supplied template
repository (so you can build your own webpage and link to App Inventor
with something like
http://ai2.appinventor.mit.edu/?repo=<myrepo>. Similarly the galleryId
value is used to point to an App in the Gallery which should be opened
upon login. The locale value determines which language App Inventor
should start in.

This commit fixes how we handle these values as we move from the client
application to the LoginServlet and back to the client application so
that they are preserved.

Note: A person’s very first login does not preserve these values because
the code that displays the terms of service fails to preserve them. This
is an old bug that we do not fix here (but can at a later date).

Change-Id: I53768cfaadd12175e3c1fe821e58024d984afcfa