Binary File Component (Extension)
This commit represents an extension. Do not build an App Inventor system
here. Instead build on master (or your favorite branch). Use this
branch/commit ONLY to build this extension.

The extension lives in the net.qyv.binfile package. Change it to
whatever you desire.

Non-visible component for reading and writing potentially binary files.
All I/O is done synchronously (on the UI Thread) and contents have to
fit into memory.  We do things this way to make it easier on the App
Inventor programmer, but it comes with a trade-off in terms of the size
of files that can be manipulated.

Build with “ant extensions”

Change-Id: I92ced8d317359e5c0bfe31428561edce7a68a29c