Fix Race Condition in Asset Loading
Asset loading is performed asynchronously and usually happens
quickly. However with many or large assets, the transferring of the
asset to the device (or emulator) may not finish in time, resulting in
an error and a re-try (which eventually succeeds).

We re-factor the asset transferring code to properly handle its
asynchronous nature using appropriate callbacks.

Change-Id: Idf0170d9a6d929a7ca964b3852811b2b09dc49ce