Cache client checks for android.keystore
Only certain actions modify the state of android.keystore on the
server. The keystore can only change if a new one is imported, the
current keystore is deleted, or when no keystore exists but the user
builds an app. However, this is not reflected in the flow control of
App Inventor. Switching between projects, for example, will check for
the state of android.keystore even though this is a user file not a
project file, and the state of the android.keystore will be invariant
between projects.

This commit caches the result of checks for android.keystore to reduce
the number of round trips to the server and only rechecks when an
action may result in an update to the status or in the event an
earlier call to hasUserFile fails.

Change-Id: I65df499d509ea7098f6b349d0da09cdaea84dd08