Add check for open project before prompting user
There is a prompt for the user to create or open a project. However,
this is based on whether currentFileEditor is set, which is a problem
is a project load is in process. This commit adds a check to
EditorManager to see if there is an editor even though
currentFileEditor is null, in which case no prompt is shown.

Change-Id: Icf044a638fef19d7fb5660a4783f852c7ccbb6cb