Add the “ShowListsAsJson” Property
When set this property causes lists to be displayed in JSON notation
rather then Lisp notation. Its default value is “false” so current
projects continue to work and new projects will still use Lisp
format. This can be changed in the future once the Companion that
supports ShowListsAsJson is widely deployed.

Change-Id: I2e6b25c32d27377685eb8111f44536479e7c55b7