Use the synchronous version of getBitmapDrawable
This reverts most of commit b5e78d7009b2066def3faae7c709c3783a17cd43.
We keep the changes to MediaUtil.getBitmapDrawable. So the actual image
fetching happens on a non-UI thread, but the UI Thread blocks until the
network I/O is done. This is non-ideal, as blocking on the UI thread can
result in an Application Not Responding (ANR) error. However this might
be more desirable then having the application apparently not load a
background image for a long time, and then having it suddenly appear
when the background network loading is finally finished.

Change-Id: I3424918a83284d19ee6ccfc9deb8824f5c0605a2