Fix image loading in getBitmapDrawableAsync
On phones with smaller memory footprints, the use of InputStream.mark
was causing an IOException to be thrown that resulted in images not
loading. This patch loads the entirety of the image into a byte[] and
then uses that in conjunction with a ByteArrayInputStream to pass the
image around to the multiple interfaces that need it without reloading
the content. This fixes the failure to load many images in the
emulator.

Change-Id: If7185e8e1a02e19a46037a4dd4df545241fa7f2b