Update OdeMessages_es_ES.properties
PitchProperties and RollProperties belong to the orientation sensor and have been wrongly translated to Tono and Lanzar in Spanish, which makes no sense in this context. They are related to orientation sensor angles, which are in turn related to navigation angles used in aircraft.

Pitch could be translated as "Tono" when related to sound, but never when relating to movement or navigation. In this case, it should be translated as "cabeceo" (meaning movement of an aircraft nose up and down, i.e. pitch) . 

Roll has been translated as "lanzar", as in "roll a dice" --> "lanzar un dado", but this is wrong in the current context. The right translation is "alabeo" (which means moving one wing up and the other one down, i.e. roll).