Fix #708 event parameter translation
When switching between languages, or loading an AIA created in one
language but the user's language is different, the translation of
event parameters would not always hold. This is due to the use of
`eventparam` as a mutation on the lexical getter and setter
blocks. Through certain actions, the `eventparam` cached value could
be reset to an incorrect state.

This commit does two things to address this issue: (1) adds an
`updateMutation()` method that can be used to recompute the value of
`eventparam` after a blocks update, including changing the dropdown or
attaching a new parent. This keeps the `eventparam` up-to-date during
normal operation (previously it was only computed on save). (2)
FieldParameterFlydown caches the untranslated `eventparam` value for
instantiating new blocks within the workspace. Previously they were
created with the value `null`, which has special meaning as a cached
'not an event' value. Now they are created with the correct mutation
at the beginning and with (1) should maintain the correct value.

Change-Id: I758b8d414d7ddf98f7375ee961187b06b47939cf