Update Backpack documentation
After the recent App Inventor 2 update, the backpack is now persistent,
which means that when right-clicking on the backpack icon on the block
editor and clicking "Show Backpack documentation", some of the text in
the dialog is incorrect. I have corrected the text to show that of the
release notes.

Change-Id: I223be859193a2b02be8bcf673bdebd6f7116444d