Update _messages.js
After the recent App Inventor 2 update, the backpack is now persistent, which means that when right-clicking on the backpack icon on the block editor and clicking "Show Backpack documentation", some of the text in the dialog is incorrect. I have corrected the text to show that of the release notes.