Add Toasts
I have added a new method called "Toast". This is different than the "Show Alert" method. It will show a 'toast' the bottom of the screen, above the device's navigation bar.

The duration can either be set to "0" or "1". 0 is Toast.LENGTH_SHORT and 1 is Toast.LENGTH_LONG. If the user sets the duration to a number that isn't 0 or 1, then it will change to 0.

https://developer.android.com/guide/topics/ui/notifiers/toasts.html