Allow component properties to specify accepted type
For the Internet of Things update, we want to allow extensions to
specify restrictions on designer properties. Specifically, we want IOT
extensions with a BluetoothDevice property to be able to restrict the
component list to the BluetoothLE extension. This commit extends the
behavior of the component property type by allowing an extension
author to include a colon (":") followed by the fully-qualified class
name of the component type to accept.

Change-Id: I6c38890fde99fe947bb1a305412d662694246463