Add support for multiple extensions per AIX
For the Internet of Things update, we want to be able to have
collections of related extensions, for example, targetting different
features of the same platform, to be imported as a single unit. This
behavior reflects how we group components such as the LEGO EV3 or NXT
robots and allows us to treat extensions in the same way.

Change-Id: I83e1069421939e57b33689f7ae0a3d5928c6c6cd