Fix logic bugs in ComponentServiceImpl and add tests
This commit fixes some logic bugs in the ComponentServiceImpl changes
for the IOT release related to the inclusion of multiple extensions
per aix and the fact that component.json was renamed to
components.json to reflect this.

This commit also adds unit tests to evaluate the different code paths
through the ComponentServiceImpl to ensure correctness of different
upgrade paths.

Change-Id: I3d8fc8998c9117297c3c8c99bede6371ef016be6