Add progress bar and _extensions in ReplMgr
This commit provides a progress bar so that the UI gives feedback to
the user when sending large assets (such as extensions) to the
Companion. This commit also adds the _extensions RPC to the Companion
to enable loading of project-specific extensions from the phone's
AppInventor directory rather than all *.jar files present.

Change-Id: I787374dcf2d7c295b2a14e38a72c6c06996ce5d1