Fix ComponentTranslation encoding on non-UTF-8-by-default platforms
Javac default output format for files is platform-specific. In the
ai.apt macro we do not specify the encoding, so we end up with
ISO-8859-1 strings on Windows that GWT later attempts to decode as
UTF-8 strings. This commit sets the encoding for the javac command to
be UTF-8 so that encodings work correctly on platforms (e.g., Windows)
where the default encoding is not UTF-8.

Change-Id: Iac68786e71983e8605e982a093d3daae09d49c01