Add “DRAINING” State to Buildserver
When running a set of buildservers behind a load balancer there is a
bunching effect that can happen when one buildserver winds up with a
large number of simultaneous builds while other buildservers are
idle. This happens when a buildserver becomes a little busy and as a
result runs slower then the other servers. This increases the likelihood
that it will get more work to do while it is still working on existing
jobs. Which slows it even further which results in more jobs until it
hits its maximum and rejects jobs (while the other buildservers are
idle!).

This change results in the buildserver returning an unhealthy status to
the load balancer (most load balancers can be configured to check a
buildservers health periodically) after it reaches 2/3 of its maximum
load. It then stays in this “DRAINING” state until the load becomes less
then 1/3 at which point the state becomes “UP” again.

Change-Id: Id6be52475151a0208000ced17fe1f2d3f15e7b94