Fix splash screen and mutators in Blockly release
This commit fixes the following:

1. Blocks in the mutator workspace did not respond to mouse down
events in Firefox versions earlier than 53.0 due to the lack of a
focus method on SVG elements.

2. Made the splash screens only appear after leaving the project list
(if the user had projects). This prevents a scrolling issue that
occurs when the user has more than one screen's height worth of
projects and the DeckPanel transitions from the project list to the
designer view while the splash screen is being shown. In some cases,
users were unable to scroll to or otherwise access the dialog to
dismiss it.

Change-Id: If368f4c15a55993f95ef17050bb8c0e9493cf40e