Make error message when loading projects with missing components
A user reported when loading an old project with a reference to the
BLE component resulted in an unexpected error. This was due to the
component being present in the Screen1.scm but not in the assets. This
commit makes it so that looking up an undefined component will result
in an error message being reported to the user indicating that the
project in question cannot be loaded.

Change-Id: Ida96802fd8152bf94dc2ac1f481c3acb3810a724