Fix redirection in Google Voice utilities
The old version of Google Voice used by the Texting component
redirects to a newer version, breaking the existing
functionality. This change uses a special URL redirection and cookie
provided by Google to legacy Google Voice to work around the
issue. This will likely break again in the future but will buy time to
implement a more robust solution, likely based on a non-Google
product.

Change-Id: Ie519a34fde18204060bd44278f206c1d4c77034b