Add failure condition when extensions collide on package name
Taifun reported that importing many extensions of the same package
name into a single project made them unable to load. This change
detects this scenario and returns an error to the user preventing them
from loading the conflicting extension. There is also a test to ensure
that if one has two extensions in the same package and then loads a
bundle containing the individually FQCN extensions the old extensions
will be removed and placed with the bundled version.

Change-Id: If35663d650d3a24f6c5aaeceaf490a24ac008522