Add support for app themes
With the addition of the Android support libraries we can now allow
users to change the theme and colors of their apps globally. This
commit adds a handful of properties to accomplish and give rough
approximations of the theme in both the web environment and
companion. Some colors cannot be changed at runtime so the only way to
see the results is to compile an app.

Change-Id: Iba0be346f1824445ab4edcdca37dd384abd42850