Make AppInvHTTPD block for REPL result
We moved the REPL operations into the main thread (unlike on Android)
and this broke the "Do It" feature. This commit makes it so that the
blocks are still executed on the main thread, but instead of
immediately returning the web server thread will block until execution
of the task completes on the main thread, at which point RetValManager
will be polled and any pending results returned to the blocks editor.

Fixes mit-cml/appinventor-sources-ios#25

Change-Id: I4d63139173205027b1c0fda2c9c1a27fbb58de81