Report java.lang.Errors to blocks editor
Stack overflow errors (and other subclasses of java.lang.Error) were
not beign reported to the blocks editor/ReplMgr, as reported on the
forum. This commit updates runtime.scm so that Throwable is caught at
key points and if the Throwable is an Error that it is converted to a
string since the dalvikvm may not provide useful information in the
Throwable's getMessage(). This resulted in an "undefined" error due to
the lack of an error message.

Fixes #909

Change-Id: I02e2065321bd92cb566f40cb478e9ec3e9c24246