Reduce number of companion updates
Opening and closing a mutator or clicking a block would result in the
blocks code being resent to the companion because any workspace event
would force a YAIL update. This change adds a isUi method to check
whether a Blockly event is a UI event, which doesn't change the
semantics of the program, and ignores such events for the purpose of
updating the companion.

Change-Id: Id80d1d094e585cc7182fe3987f8e0adc2b45b872