Make NetworkUtil IP address detection more robust
Sometimes when running the companion in the simulator the IP address
of the computer is reported as "error" rather than an available
network connection. We are looking only at `en0`, which should be the
default on most devices, but on some Macs the Airport is `en1`, which
breaks our code. This implementation is more lenient with respect to
the name of the network interface.

Fixes mit-cml/appinventor-sources-ios#45

Change-Id: I9bdf25e570927d672d070d8e17d50960dce6aa3e