Make if block mutator robust to undo/redo operations
This commit reconciles our implementation of the `if` block with the
implementation in Blockly. In particular, our implementation only
works under the assumption that the user is making changes via the
mutator and causes graphical problems when the undo/redo feature in
Blockly is used. The Blockly version correctly handles undo/redo, but
the mutator is slightly different in its implementation. This change
maintains our visual presentation of the mutator but uses the Blockly
logic for updating.

Fixes #973

Change-Id: Ib335d4cd7f132f8bf21f9c691ee7a7e53f28e674