Add IgnoreSslErrors to Web and improve error messages
This change adds an IgnoreSslErrors to the Web component that works
similarly to the property of the same name on WebViewer. Enabling this
property will suppress hostname and certificate chain validation for
any operations with the given component. There is also a new error
message for reporting SSL errors separately from Web errors that gives
more detailed feedback about the SSL handshaking procedure (useful for
debugging certificate issues on the Companion).

Fixes #791
Fixes #886

Change-Id: I595892920e92543bfaea590bdaa86e0caf781563