Defer loading of screens until after Screen1 loads
When importing a project that requires an upgrade involving project
settings, if any screen loaded before Screen1 would fail to
upgrade. Attempting to build a project in this state would only
generate YAIL files for the successfully loaded screens. The build
server would attempt to generate YAIL for the missing screens using
the AI1 YAIL generator, which no longer exists. The error message
doesn't give any indication that this is the problem.

This change defers loading of screens other than Screen1 until
Screen1's form is loaded. At this point the upgrade of the Form and
the project settings will have occurred, and then we can proceed to
load the additional screens.

Change-Id: Ibde055171a8939a9e9ae5ed8ea5ce3954fb5ccf1