Update whitelist.py and /remote_api
Update whitelist.py so it works with the latest App Engine SDK. Also
update permissions on /remote_api to be compatible with the latest App
Engine SDK. Note: The change appears to remove the requirement for admin
rights to use the /remote_api facility. However this admin requirement
is now enforced by the Google supplied code which runs behind
/remote_api. Performing the additional check in web.xml just causes an
error about too many authentication attempts.

Also remove excess catch statement in Whitelist.java that has been
giving us compilation warnings for a very long time. Note:
Whitelist.java is no longer in use, but can be enabled via a simple edit
to OdeAuthFilter.java, so we are not removing it yet.

Change-Id: I30cbb0eba3fba3f20d8443a94dae837c827f0b20