Add “Shared Backpack”
A shared backpack is enabled at SSO login time[1] by providing a
backPackId parameter in the SSO Token. Once enabled backpack contents
are not stored with the user account but instead in a specific location
defined by the backPackId. Multiple users can use the same shared
backpack at the same time. Each time the shared backpack is opened a
fresh copy is loaded from the server.

There is still a race condition if two users attempt to store blocks at
the same time. One will win and the other’s changes will not be added to
the backpack. This may be corrected or mitigated in a future release.

[1] Note: SSO Login is not currently implemented in the App Engine
version of MIT App Inventor. This commit updates the Backpack code to be
in alignment with the code we use in our Hong Kong deployment which is
not run in App Engine. We plan to implement SSO login in the App Engine
version in the near future.

Change-Id: I5957a42999ff14d7263a47993efaf8f3492c016g