Add the "Break" block (#1180)
This adds a break block that can be used to break out of for-range,
for-each, and while loops.  Running excuting break terminate execution
of the innermost loop.  It is an error to have a break block that is
not contained in a loop.  Note that this change does not implement
continue, only break.

Change-Id: Ib480b03d0db70d85016a7400579b8fe14435b12a