Added LoadingDialogMessage and ShowLoadingDialog properties to Fustiontables component
This commit adds two new properties to the FusionTables component. The first property is "LoadingDialogMessage". This allows the user to customise the default "Please wait loading..." text which is shown in the dialog after calling the .SendQuery function.
The second property is "ShowLoadingDialog" this allows the user to choose whether or not to show the dialog.