Update How Build Server progress is reported.
Buildserver progress, as reported in the progress bar during a build,
used to work by having the client poll the App Inventor server, which
would in turn poll the buildserver to learn the progress of the build.

However this code would only work properly if there was only one
buildserver and only one build in progress. In reality we handle dozens
of builds simultaneously via a number of independent buildservers. So
the progress being returned to the user was the progress of *some* build
in the system, most likely not theirs!

The constant polling by the App Inventor server also causes problems
with the load balancer in front of the buildservers. This is because
most “jobs” are not real build jobs, but just polls for progress. This
may result in a non-optimal distribution of jobs between the various
buildservers.

This code changes things so that the buildserver does a callback to the
App Inventor server when appropriate milestones are reached in the
build job. The client still polls to App Inventor server for progress,
but the App Inventor server no longer polls the buildserver. This should
result in a better distribution of build jobs on the buildservers and
also returns to the end user the correct status of their build.

We also removed some obsolete code that attempts to generate the Yail
within the Buildserver. We *always* generate the yail in the client now
and it is sent to the buildserver.

Change-Id: Ibaa9a7e3ff6576335d5ab6261a4121f27a1daf1d