Enforce that block definitions exist for designer properties
If a designer property is included in a component without a
corresponding block entry, then the YAIL code generator will throw an
exception attempting to look up the type of the property when
generating set-and-coerce-property! code for the designer. This change
enforces that every @DesignerProperty have a corresponding
@SimpleProperty to ensure that components and extensions include the
necessary information.

Change-Id: I85cf53b1b657254658fd0cb0f05af3dde6b32740