Implement custom zoom controls
The Android SDK provided zoom controls used by osmdroid are not scaled
by ScaledFrameLayout, and so do not work correctly when Sizing is
Fixed. Further, the ZoomButtonsController is deprecated starting with
SDK 26 and Google recommends that developers provide custom views for
zoom controls.

This change implements zoom controls for the Map component similar to
those shown in the designer.

Change-Id: Ifd85de7fc60a042eb1cddc23c6b037964395da7d