Close blocks drawer when deleting component (#1249)
If one deleted a component while on the blocks screen, the drawer
remained open. This commit tracks the last component opened and, if
the deleted component name matches the last component opened, it will
close the drawer. We track the last component referenced so that we
don't hide the drawer in the event a component is deleted when another
drawer, for example the built-in math blocks, is open.

Change-Id: I4808a069b38986e44892e721b2d35132bfdae943