DEBUG: Add code to enumerate permissions
Add code to enumerate requested permissions in the manifest. We can use
this information (in code yet to be written) to decide whether or not we
need to ask for access. At the moment WRITE_EXTERNAL_STORAGE may fit the
bill because there are lot of places where deep in the call stack we may
discover we need it.

Change-Id: I927bc15c43935c0801fd2f4b7fe51e27bf000b60