Ask for Write External Storage permission at start
If an app has WRITE_EXTERNAL_STORAGE permission declared in its
manifest, then request permission from the user as soon as the
application starts. We do this because a lot of the times when we
dynamically discover that we need permission, we are deep in the stack
and not in a good place to recover. Or at least I believe that is true.

We will continue to look for a more granular solution.

Change-Id: I36b6d136b7aab3d64da73b3fdb7ed5a533200178