Add screenSize config flag to handle Form rotation
Starting with API level 13 (Android 3.2), there is a new screenSize
change that can be sent to activities. This is in addition to the
orientation change App Inventor already handles. If not appropriately
handled, Android will attempt to restart the activity, which "crashes"
the companion. This change informs Android that it should not restart
an App Inventor app if it gets a screenSize change.

Change-Id: If3f00bc495cd0dc2cd49378a6b5f145994ce4175