Add Online Documentation for Built-In Blocks to repo
The advantage to tracking the documentation in the same repo as the code itself is that changes to docco can be tightly coupled to the code changes it describes.

blocks.html has been changed to index.html. Images are also added to the repository with related links updated. All other links in the html are now external links appinventor.mit.edu/explore