Fix theme handling with new AppCompat library
With the SDK 26 update, a hack we used to infer the theme was
broken. This change properly implements a mechanism to handle the
theme support for older devices by passing a flag at codegen/build
time rather than attempting to infer it from the SDK behavior on the
device. It introduces new theme helper objects that can be tailored to
behavior specific to an Android version.

Change-Id: I7d6e993cc7a641b6242789c26c0cc62c4be61a3f