Trim extraneous whitespace from base64-encoded aia
The base64 utility on macOS writes a carriage return at the end of the
base64 encoded string. When App Inventor attempts to read a project
template with this extra whitespace, it causes an exception to be
thrown since the server's decode() implementation expects a string
with a length multiple of 4. This change switches to using
decodeLines(), which ignores whitespace.

Change-Id: I34f6c7b1ef1fcd12b82eab67c98ffb422bbf5894