Remove BOM from GeoJSON files
While GeoJSON files should not have a BOM per the spec, sometimes
systems will include the UTF byte order marker (BOM). This breaks the
JSON parsing on both in the web client and app. This change will check
for and strip out the BOM before passing the JSON string along for
further processing.

Change-Id: I9452871e19b3985f9beea48b30cf49b60e4835cb