Fire permission granted without posting runnable if we have permission
Clocks break on SDK 26 due to an order of operations problem that is
introduced by asking for dangerous permissions. The traditional order
of operations would be to call $define() as part of onCreate and
therefore before onResume. With the mechanism for requesting dangerous
permissions, we can end up calling $define before onResume if the
dialog is shown in response to needing SD card access, or after
onResume if we need SD card access but the user has already granted
it. In the latter case, this causes the Clock's onScreen field to be
false and prevents the timer from firing. This commit moves the check
for permission to be before posting the runnable so that if we know
permission is grantd the old control flow remains intact.

Change-Id: I01330f8cc373e1c5184dedc7eb208199b74fe8fb