Make ScrollView fill viewport on Nougat and higher
The semantics of how ScrollView manages its content changed between
SDK 23 and SDK 24. Prior to SDK 24, the content of the ScrollView
would stretch the content bounds of the ScrollView, which resulted in
the height fitting the content. With SDK 24, this relationship isn't
preserved, so children that are supposed to fill the parent no longer
do so or take on a different size.

This change sets a flag so that the ScrollView will allow its content
to fill the viewport on SDK 24 or higher. We retain the old behavior
on Android versions older than 24 to minimize breakage.

Change-Id: I0a4c1f1588cc54a499c5860794fb77ad83fa918d