Autogenerate OdeMessages file for translatable keys
When adding a new user-facing API, an entry needs to be added to
OdeMessages for translations. This is particularly cumbersome for
people developing extensions because the changes to OdeMessages that
need to be made won't go into the system. This change makes it so that
the component description generator will also generate an
AutomatedOdeMessages file with all of the necessary entries, no longer
requiring the developer to add their own entries.

Change-Id: Iff1fd1307cee03d0b71d8a4921f52f469661df3d