Enable debug panel logging only when admin
Projects with many components may cause a large number calls to
OdeLog.log, which appends messages by converting the debugger panel's
content to HTML and back. This change enables logging only if the user
is an Admin, since only Admins have access to the debugger panel to
begin with. This reduces the number of cycles spent on logging for the
vast majority of App Inventor users. We also optimize how the log is
constructed so that users with Admin rights have reasonable
performance on complicated projects.

Change-Id: I62cc2af0e421d3f5fda668f72de62c1895414092