Implement versionName and dateBuilt annotation fields
This change adds two new fields toe @DesignerComponent. The first is a
versionName field so that extensions can display custom version names
rather than the version number. This can be used to mark an extension
build as beta or a release candidate, for example. The second change
is a dateBuilt field that is populated by ComponentProcessor at
compile time. This is useful for knowing when an extension is
compiled. Both pieces of information appear in the component help
widget, and version name is preferred over version if provided.

Change-Id: I12c97c9c0a3b18abd64a539f1c5072b975d47d14