Save and restore user preferred locale
For non-English users of App Inventor, if they go to the main page or
click a link without a locale specified, for example, a repo link,
then they will be presented App Inventor in English. This is bad from
a UX perspective as the user then has to change the language and wait
for the site to reload. It also interrupts whatever workflow they were
currently doing (e.g., when opening a template project).

This change stores the last locale from the query string as the user's
preferred locale. When a locale isn't specified in the URL, we will
check the locale and if it is set, redirect to that page
automatically. To also save on performing actions that would be
canceled by the redirect, we also reorder some initialization of Ode
so that it only occurs if the redirect won't happen.

Change-Id: I1b9ffa756aa08f05495832768b242341e4a30c38