Fix error 1101 on Android versions before Nougat
Installing a package via the QR code scanner fails on Android versions
before SDK 24. The issue is that we switched to using content: URIs,
which is required on SDK 24+, but this fails to resolve the package
installer activity on SDK < 24. This commit adds a helper method for
constructing the correct URI based on the device's SDK version.

Change-Id: Ieaec37b79d6189f75535b5a9be4d0f38b756e63a