Implement copy and paste support in the designer
This change adds support for copying and pasting components in the
designer. It handles the entire tree from a given component (and
supports multiple roots for if we allow multiple selection in the
future). Name substitution will occur if a component exists with the
same name when pasting. Top level blocks referencing the copied
components are also included in the operation. Holding the shift key
while pasting will disable pasting the event handlers.

Resolves #726
Resolves #840

Change-Id: Ica9637f8d07e7287cd698248c60533a5e10af752