Simplify the WebRTC Poller
Eliminate the single threaded executor because it is not needed. Instead
use Timer.scheduleAtFixedRate() to trigger the Poller once per
second. This approach is more correct and simpler.

Also, ignore an empty response from the Rendezvous server (which means
that it hasn’t received any ICE Candidates from the browser
yet). Previously we attempted to JSON parse it and threw an
exception (which was logged and ignored).

Change-Id: If0bc95c754a3fd052ac32cfa966fcbcfe658f55d