Fix logic bug when property value is the empty string
The conditional properties change added the ability to use a default
value when a property isn't specified. However, there is a logic bug
due to the fact that the empty string is falsey in JavaScript. This
fix also checks specifically for the empty string as well so that the
default value will be sent.