Fix theme issue due to property/component initialization
The change to make component reference blocks usable in globals
changed the ordering of initialization in such a way that components
that rely on the theme applied to Screen1 end up getting the wrong
theme. This corrects the order of initialization so that Screen1's
properties are set immediately before any other components are
created.