Handle NSNumber-wrapped bool in YAIL-ObjC bridge
Booleans are stored in NSNumber as 0/1 with the Objective-C type of
signed char. The YAIL-ObjC bridge code treats signed chars like any
other integer and upcasted to int values for use in the Scheme
code. This results in boolean values from Swift being converted into
0/1 in the Scheme code. This change interprets the signed char values
as true and false so that booleans are passed correctly. This also
means that we should never use signed chars for representing integers
because they will be downcast to booleans.

Change-Id: Ifc46d7f463c0d7de0c2e6ca93bc57e382346a9d0