Implement better handling for large integers
In the Objective-C to YAIL bridge, an NSUInteger or NSInteger would be
cast to an int. However, on iOS these types are 64-bit, not 32-bit, so
if the value was outside the range of 32-bit int, it would get capped
at the min/max value. This change makes it so that if the value is out
of range of int type it will be returned as a double precision
floating point value. Longer term we will need to pull in a bignum
library in order to support larger integer values due to picrin's use
of NaN boxing.

Change-Id: Ib6a45f98b9c79ed7a37040ac0007865b5b7e0a32