Mitigate WebRTC Issue
The previous commit updates the Companion to better handle the case
where the Rendezvous server is far away from the user (long round trip
time).

This commit does a mitigation, at the cost of 5 seconds, in replmgr.js
that results in older Companions being more likely to connect with a
long round trip time.

Change-Id: Ice6b86b11a3c21beedf2c066ac46593659651713