Blocks Toolkit (#1728)
Feature adapted from Lynda Tang's graduate work.

Adds project-level property "BlocksToolkit" (internally also called
SubsetJSON), which is a JSON string describing UI components and code
blocks to be displayed in the Designer and Blocks Editor. This allows
projects to limit the options in the GUI to fit the curriculum and not
overwhelm students.

Change-Id: Ie37ecbaa31398fcfcc57320a01dd1f67d1b13522