Register YaFormEditor and YaBlocksEditor earlier
Prior to this change, YaFormEditor and YaBlocksEditor are registered
in loadDesigner, which is called from onShow. This is problematic
because if the screen hasn't been opened in the current session and
the user deletes an extension, the unopened screens are not notified
of the change. When the project is reloaded, it will fail to load
because the unopened screens still referenced the deleted extension.

This change moves the registration of YaFormEditor and YaBlocksEditor
as FormChangeListeners to the onFileLoaded method so that they are
made aware of project-level changes (like removing an extension) even
if they are never viewed in a session.

Change-Id: I8306ede85f5fb3576d810c9bd213e5e00df938ff