Add %type% as a token in component descriptions
App Inventor doesn't present a type hierarchy even though internally
it makes sense for objects to have shared methods defined in super
classes. However, the documentation for the properties, methods, and
events of an App Inventor component shared in this way must typically
be written generic enough so that they make sense when shown on
concrete classes. This can lead to either awkward/generic sounding
documentation or requires that the component author override all of
the relevant features on the component to provide component-specific
descriptions.

This change makes it so that descriptions of components, properties,
methods, and events can contain one or more %type% tokens that can be
substituted with the name of the concrete representation shown to the
App Inventor user. This allows for documentation to be shared at the
super class level while still making it sound specific to the
individual component type.

Change-Id: Ida4eb22bec12b2e33ac73de2bb5706d1683b227a