Make AWT run in headless mode in buildserver
On macOS, the application icon preparation step causes Java AWT to
create an app window. This window gains focus, even though the
buildserver is a background process. This really only affects
developers working with the sources on OS X. This change sets a flag
to inform AWT to run in headless mode, which prevents it from creating
the app window.

Change-Id: Ida725b80e67c55777437cdd69392a3fab4dcf00a