Fix empty point, holepoints after polygon drag
This one is a mea cupla. When I implemented dragging of polygons,
there only a single polygon was allowed (and I called
getPoints/getHoles to access its data). However, when we started
loading more complex forms using GeoJSON (e.g., the US states map), we
needed to support MultiPolygons, so I implemented a class to do
so. This class was really a proxy class that grouped calls to its
underlying children. The issue here is that its getPoints and getHoles
methods then report empty lists because it holds no state. This change
adds methods getMultiPoints and getMultiHoles to access its
content. Different methods are needed because the Java language
doesn't allow the same method to return different types before type
erasure.

Change-Id: Ic56b10d6ca1df949a2bbf342b4bc7d3d473a6342