Clone default button background drawable if needed
When we apply colors to buttons, we apply a filter to the background
drawable. However, the default drawable is cached and reused across
all buttons. This causes problems when mixing buttons of the different
colors with buttons using the "default" color.

This change addresses the problem by making a copy of the background
drawable and applying the color to the copy so that the currently
rendered button doesn't adversely affect buttons that come after it in
the view hierarchy. This copy is done lazily at the time the color is
changed from "Default" to something else so that we can still benefit
from the reusability of the default button background drawable.

Change-Id: I2be3b15df32b76a59a29f1e2aed41054f5bdd975