Send code after asset refresh, not before
If someone changes the image of a button by using the upload file
feature, we will send the code with the new image property before the
asset is sent to the device. This results in the button not getting
the appropriate image. This change moves the operation to build the
YAIL into the callback of refreshAssets. It also makes sure that the
callback is called regardless of whether there are new assets to be
sent so that YAIL generation doesn't break in the absense of assets.

Change-Id: I14520d8a7945902aa4a9b345d3f9d164a3a9448e