The New Gallery
This commit adds the interface support for the new gallery.

The new gallery is a separate system run on its own
infrastructure (which today consists of a horizontally scale-able
collection of Linux virtual servers).

It is configured via three system properties (out of
appengine-web.xml). They are:

gallery.enabled –  Set this to “true” to enable the Gallery
gallery.id –       This is a string which is used to identify this
                   instance of App Inventor to the gallery. It is
                   also used to identify the shared keys needed
                   to authenticate tokens sent to the Gallery
gallery.location – The location (URL) of the gallery itself.

Change-Id: I30745a65c976cadd288650f779a9bf546097567a