Add ant flag force.builddocs
This change adds a property force.builddocs to the build script that
controls whether a failure to successfully execute Jekyll constitutes
a build failure or not. It defaults to false. In Jenkins we will set
it to true so that builds fail if they cause jekyll to fail. This way
people who do not have Jekyll installed can continue to work with the
sources. Since everything needs to pass Jenkins before being merged,
maintainers can run Jekyll to update the docs prior to merging.

Change-Id: I9fb22eea7582fb5ba5a83069778c23a40c14aaa8