Add TODO tags on loop-block-related logic
I think that some of the lgoic I have tagged could be refactored to
call functions that exist on the block. This would make it easier to
maintain, and possibly more readable.

Though I'm not sure how feasible these TODOs are, specifically because
of how they might affect extensions, which I know nothing about.