Another fix for WebRTC
Cleanup the code that deals with detecting connection failures. This
change results in a more stable experience on Firefox. It also ensures
that any closed WebRTC connection is cleaned up properly. This means
that if the Companion disconnects due to the Companion being closed, or
a network failure, a new connection can be made reliably without having
to reload App Inventor in the browser.

Change-Id: I35ccba26729d7c870201165cb3821991a6f523ee