Implement preventScroll for focusable elements
The source structure tree is based on GWT's Tree implementation, which
uses focusable elements to handle keyboard navigation. However, when
these items are selected by the user, they also cause the browser to
scroll, resulting in the view jumping around, sometimes significantly
depending on the size of the scroll offset. This change passes an
optional flag, `preventScroll`, to the native `focus` method to
suppress this browser behavior. It leverage's GWT's mechanism for
making compile-time substitutions of classes to replace the GWT
implementation with an App Inventor-specific implementation.

Change-Id: I4766e5a3c119ee461d8f20dc359f4d9cdec112c0