De-bounce the Publish to Gallery button
Some people were double-clicking this button which results in two copies
of the project being sent to the Gallery simultaneously. This results in
two entries for the project appearing in the Gallery. Then when the
author next goes to update it, the Gallery finds two copies and fails.

We will be adding a uniqueness constraint to the Gallery database which
will prevent this. However this will result in one of the two
transactions failing. Depending on the exact timing of things this can
still result in the end-user receiving an error.

This change “de-bounces” the Publish button so while one button push is
still being processed, a second is ignored silently.

Change-Id: Idd7506773b5e2a27500250d82a1350602b634a77