Add “UpdateDone” event to CloudDB
The UpdateDone event means that a store operation (StoreValue,
AppendValueToList or ClearTags) has completed. Because some
applications, such as SketchAndGuess can cause numerous store operations
in a short period of time, there is a risk that these events may
overload the sending device. In the case of a large number of update
events, only one UpdateDone event will be triggered per tag updated
once the backlog of update operations is cleared.

The most likely use case for UpdateDone is to handle storage of large
objects. All store operations are performed in the background, so this
event permits apps to know for sure when an update operation is
complete.

Change-Id: Ia6ff5760ad622efecc9a8d235a177b9c3e00c6a9