Properly handle error while saving files
When an error occurs while saving a file we need to put the associated
file editor back in the dirty editors list. Otherwise data may be
lost. Prior to this fix once i/o failed, the editor was considered clean
and i/o would not be retried unless the editor was modified again. So
for example if a designer attribute was changed, but the i/o failed, it
would not get saved again, even if a change was made in the blocks
editor and even if the “Save project” menu was used.

Note: We do not re-schedule the failed i/o but instead depend on another
user change (anywhere) or explicit use of the “Save project” menu. I
considered re-scheduling the i/o but am concerned that if i/o is failing
due to a server problem, a lot of clients retrying i/o may make it worse
via the “thundering herd!”

Change-Id: Ife50c3c20d407b4e009c639bc36c2596331342d7