Restore TextSize() to the ListView Component
The recent rework of the ListView component removed the TextSize()
property and replaced it with the FontSize() property, which takes a
float instead of an integer.

However this breaks backwards compatibility. So this change restores
the TextSize property, making it an interface to the FontSize
property. I.e., it converts its input to a float and calls FontSize()

The FontSize property and block are currently hidden to avoid problems
with older Companions. At a future date it can be un-hidden.

Change-Id: I6f177645a74d914d15da73bffb31c5f68a3f1086