Remove Blobstore Conversion
Years ago MIT App Inventor stored "large" objects in the now
deprecated Google Blobstore. In 2015 we added code to move objects
from the Blobstore to the new Google Cloud Storage (GCS) system. At
first this was done on demand and eventually we ran a map-reduce job
to convert remaining objects. Not long after Google disabled the API
we used to access the Blobstore.

This commit removes the code for that long finished conversion. Note:
There is still code to read objects that may be orphaned in the
blobstore. However we always write objects to either the datastore or
to GCS. It isn't clear if this code to read from the blobstore still
works... but better to leave it for now.

Change-Id: Ie28102e61d5836301a7fee63eb837dbe3fe0c62b