Use Authkey in sources if there is one
Normally there is no authkey stored in the source tree. Instead
developers run "ant MakeAuthKey" to generate a personal authkey which
is then copied into the build tree.

However our branding branches include an authkey in the sources (in
war/WEB-INF), often customized for the installation. This change
causes this source resident key to over-ride the developer's personal
authkey.

Change-Id: I47761db8028d2432b3f5f0dcb1ece725ef5af0ac