Provide a new "Delete Account" Menu Item
This action will remove an account provided that there are no
non-trashed projects in the account. If there are, a dialog is
produced explaining that all projects must be delete first.

If there are no projects, then the remaining vestiges of the account
are removed, along with any projects that are marked as trash.

If a deleteaccount.url property is defined in appengine-web.xml, then
upon account deletion, the user will be redirected to it with a
"token" parameter set to an encrypted protocol buffer with the
information the login server needs in order to cleanup what remains of
the account there.

Change-Id: I519e973fb4fa4362a40f680012360badec2128a4