Load Assets directly from the App Inventor Server in legacy mode
This code changes how assets are loaded in legacy mode to use the same
strategy as WebRTC. Namely, tell the Companion to fetch the asset
directly from the App Inventor server instead of first loading
it (inefficiently!) into the browser and then stuffing it into the
Companion. This does mean that the Companion needs to be able to reach
the App Inventor server.

Note: There is some refactoring of how forms are pushed to the
Companion, these are in effect fixes for race conditions in the legacy
code that were not apparent until we used it for loading assets. The
code to use an HTTP request to learn the Companion version has also
been removed. This is now done via the Rendezvous server, so this code
is obsolete and has not been used for quite some time.

Change-Id: Ida31a8b1373f7a76a5da1f79cc2f4bd9a2b8afc3