Fix: do not read in whole file in UploadServlet
The change is actually in FileImporterImpl.java. Previous code read
the whole object into memory and then threw an error if it was too
large. This can result in a Java Heap overflow, which is bad. So we
incrementally check the incoming data size as we read it. If it is
larger then the configured maximum, we keep reading, but throw the
byte away instead of consuming memory.

Change-Id: Ife3180eacb4ec5037be077142849667d7010b102