Report the contents of /etc/hostname in buildserver "vars"
We add a new variable "etc-hostname" which contains the contents of
the /etc/hostname file.

Within a docker swarm environment we can bind mount /etc/hostname in
the container running the buildserver to the /etc/hostname file on the
actual host the container is running on.

Today when running multiple buildserver containers, the "hostname"
variable contains a string of random characters and digits, which
isn't helpful in locating the actual host (or VM) that the container
is running on. It is sometimes useful to know this!

The "hostname" variable itself is retrieved by calling
InetAddress.getLocalHost.getHostName() which does not use
/etc/hostname. So we add this new variable which just reports out the
first line of /etc/hostname, which if bind mounted, will be the
hostname of the actual VM or host.

Change-Id: Ib9d495f52e4bcae7bdd2bf1413ba9a9fd666e265