Move Resources back to using static map
Earlier versions of the build server used a shared map of resources
names to files on disk. Those files only ever got deleted when the
server was shut down and otherwise were just reused between
builds. After refactoring, the map was changed to a per build
resource, but the resulting files were not appropriately cleaned up at
the end of a build. This would eventually lead to the buildserver
running out of disk space and silently failing builds.

This commit reverts the semantics of the resource map to be a single
shared resource across all build threads, with the resources only
being cleaned up at server shutdown.

Change-Id: I374e88c5a08a81e7904f072b42b89f4468c9f646