Fix checkDropDownContainsValidValue
We can remove the drag test since it's also done elsewhere, in a better way.  More importantly, this was also preventing the dropdown check from occurring when it needed to.  Finally, note that this wasn't even properly testing whether the block was being dragged because the neither of the properties that it was comparing exist in Blockly anymore, so the comparison of undefined to undefined was giving us a misleading, and incorrect, value of true!