Add support for input image to ChatBot protocol buffer
I have added the protocol buffer changes to input an image to
ChatBot. I have updated performRequest to take a third argument, a
bitmap, which may be null if there is not input image.

NOT FINISHED

TBD: Convert the source image in the block into an appropriate
BitMap. Hint: Look how ImageBot does it. That code should work here as
well.

Note: The production ChatBot proxy will ignore the input image as of
this writing, but I (Jeff) am working on adding that support.

Change-Id: Iedfbee9dca0cb7a072b6a9fd9c13be662bb89df4