Create custom connection checker
This connection checker allows functions as types, and calls them in order to allow more flexible type checking on a per connection basis.

Note that this already existed in a different form, via some monkeypatching that we did, but now can be done via a custom connection checker, now that that capability is supported by Blockly.