Fix extension loading for Android versions < 14
If we are on Android 14 or higher, the final jar file's name is based
on the name of the extension.

Older versions just name it classes.jar

Make sure we can write the classes file before we attempt it.

In Android 14+ we set extension class files to be read-only. We need
to write-enable the file before we attempt to over-write it with a new
version.

Change-Id: I4295815718eaeb91d504b478c6e44f695dcac488