Redirect to the login page directly from index.jsp
When the index page was a straight HTML page, we would load it. It
would then load Ode, which is not small! Ode would then check if you
are logged in and if not redirect you to the login page. The login
page would then redirect you back to the index page once you are
logged in. This would now load Ode a second time which would again
check if you are logged in. This time you would be logged in and
project loading etc. would continue.

The index page is now a jsp, so we can execute code in it. This change
causes index.jsp to do the login check and if you are not logged in,
it redirects you immediately to the login page rather then loading
Ode.

Ode still does the login check, but it should now always be true. I
have intentionally left that check in place. It isn't expensive and we
may have situations where having the check there makes sense.

Change-Id: I6c7f220a32d2734f1ce3242a01a0b2bcf45fd596