Add Image Creation to ChatBot
Add a “CreateImage” and “GotResponseWithImage” blocks to ChatBot.

As of this writing, we only support using Gemini to create
images. Because Google is updating models frequently, we recommend
that people using ChatBot for image generation set their “provider” to
“Gemini” but leave the model field blank, selecting the default.

The ChatBot proxy will then select the appropriate model for image
generation.

Change-Id: I8a444c36285e82298128b41d349a59eeab3775d9