Add Image Creation to ChatBot (#3462)
* Add Image Creation to ChatBot

Add a “CreateImage” and “GotResponseWithImage” blocks to ChatBot.

As of this writing, we only support using Gemini to create
images. Because Google is updating models frequently, we recommend
that people using ChatBot for image generation set their “provider” to
“Gemini” but leave the model field blank, selecting the default.

The ChatBot proxy will then select the appropriate model for image
generation.

Change-Id: I8a444c36285e82298128b41d349a59eeab3775d9

* Fixes and refactoring based on review comments

Change-Id: Ie2f526827ac60438f31ef195704207e10c6a770d

* Always trigger GotResponseWithImage when doImage is true

Change-Id: I078aebe1c99207116e145173b5fa06883af78e5e

* Rename parameters to be congruent with ImageBot

Change-Id: Ic1a7a3bc16ae601ebe9ebbf81298dbd61991085e