Fix Header Alignment on AdminUserList Page
When I originally built the AdminUserList page I was lazy and re-used
the CSS for the ProjectList. However when the new UI binder code was
introduced, the ProjectList CSS was updated, which broke the
AdminUserList page, which was not converted.

This change just gives the AdminUserList page its own CSS which works
properly.

Change-Id: Ic4f035b2f0ec2b2174810293a03cf604f9cccbc8