Verify locale
Verify that the locale being provided in the URL is a valid locale and
if not, default back to English. This fixes the issue where a nonsense
locale results in a blank screen, which is sticky. I.e., the bogus
locale was stored in the user settings before it was known to be
bogus. This resulted in people continuing to get a blank screen when
they went to App Inventor without specifying a locale, because the
bogus locale stored in their user settings was being attempted.

Change-Id: I42e842ac6c1455e3fc02c0e1dc3fa1ca3fd7a36e